Move subnet allocator from pkg/util/netutils to pkg/network/master
- Subnet allocator is specific to SDN master and not used anywhere.
