+++ pkg/network/master/subnet_allocator.go
@@ -1,6 +1,7 @@
-package netutils
+package master
 
 import (
+	"encoding/binary"
 	"fmt"
 	"net"
 	"sync"
@@ -132,4 +133,14 @@ func (sna *SubnetAllocator) ReleaseNetwork(ipnet *net.IPNet) error {
 	sna.allocMap[ipnetStr] = false
 
 	return nil
+}
+
+func IPToUint32(ip net.IP) uint32 {
+	return binary.BigEndian.Uint32(ip.To4())
+}
+
+func Uint32ToIP(u uint32) net.IP {
+	ip := make([]byte, 4)
+	binary.BigEndian.PutUint32(ip, u)
+	return net.IPv4(ip[0], ip[1], ip[2], ip[3])
 }
