+++ pkg/network/master/subnet_allocator_test.go
@@ -1,4 +1,4 @@
-package netutils
+package master
 
 import (
 	"fmt"
@@ -246,5 +246,21 @@ func TestAllocateReleaseSubnet(t *testing.T) {
 	sn, err = sna.GetNetwork()
 	if err == nil {
 		t.Fatalf("Unexpectedly succeeded in getting network (sn=%s)", sn.String())
+	}
+}
+
+func TestConversion(t *testing.T) {
+	ip := net.ParseIP("10.1.2.3")
+	if ip == nil {
+		t.Fatal("Failed to parse IP")
+	}
+
+	u := IPToUint32(ip)
+	t.Log(u)
+	ip2 := Uint32ToIP(u)
+	t.Log(ip2)
+
+	if !ip2.Equal(ip) {
+		t.Fatal("Conversion back and forth failed")
 	}
 }
